requirements.txt
polars==0.20.13
beautifulsoup4==4.9.3
requests==2.31.0
pyyaml==6.0.1